requirements.txt
asgiref==3.5.2
Django==4.0.8
django-appconf==1.0.5
django-compressor==4.1
django-environ==0.9.0
mssql-django==1.1.3
pyodbc==4.0.35
python-slugify==7.0.0
pytz==2022.6
rcssmin==1.1.0
rjsmin==1.2.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
